feat: support writing file contents directly from SPC not in secret manager
A possible implementation/fix for:

https://github.com/kubernetes-sigs/secrets-store-csi-driver/issues/948

This would allow the GCP secret provider class to write some static
non-secret data to a file next to the actual secret. This type of
behavior is needed because some systems like ArgoCD take in all
configuration from a single K8s secret with different key/values.

This could also be used to put public keys (and keep the private key
as just the secret value).